requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
